Relocate conflitc.tsx under components/PageEditor
